chore: convert consuming configs to TypeScript
Replace eslint.config.js and commitlint.config.js (root + all packages)
with .ts equivalents. Type each with satisfies (Linter.Config[],
UserConfig) for editor-level type safety without a build step. Add
@commitlint/types as an explicit devDep so the UserConfig import
resolves correctly.
