--- commitlint.config.js
@@ -1,9 +1,0 @@
-/**
- * @see https://commitlint.js.org/reference/configuration.html
- * @type {import("@commitlint/types").UserConfig}
- */
-const config = {
-	extends: ["@theholocron"],
-};
-
-export default config;
