+++ commitlint.config.ts
@@ -1,0 +1,7 @@
+import type { UserConfig } from "@commitlint/types";
+
+const config = {
+	extends: ["@theholocron"],
+} satisfies UserConfig;
+
+export default config;
