+++ eslint.config.ts
@@ -1,6 +1,6 @@
+import type { Linter } from "eslint";
 import { library } from "@theholocron/eslint-config/bundles/library";
 
-/** @type {import("eslint").Linter.Config} */
 const config = [
 	...library(),
 	{
@@ -11,7 +11,13 @@ const config = [
 			"n/no-unpublished-import": "off",
 		},
 	},
-	{ ignores: ["dist/**", "coverage/**"] },
-];
+	{
+		ignores: [
+			"packages/*/dist/**",
+			"packages/*/coverage/**",
+			"**/node_modules/**",
+		],
+	},
+] satisfies Linter.Config[];
 
 export default config;
