+++ package.json
@@ -25,6 +25,7 @@
     "@theholocron/cli": "catalog:holocron",
     "@theholocron/holocron-plugin-github": "catalog:holocron",
     "@commitlint/config-conventional": "^19.8.1",
+    "@commitlint/types": "^19.8.1",
     "@semantic-release/changelog": "^6.0.3",
     "@semantic-release/exec": "^7.1.0",
     "@semantic-release/git": "^10.0.1",
